+++ Formation/Common/Field/Field.php
@@ -497,6 +497,18 @@ class Field {
 
 		$checkbox_radio = 'checkbox' === $type || 'radio' === $type || 'radio-select' === $type || 'radio-text' === $type;
 
+		/* Radio text size */
+
+		$radio_text_size = '';
+
+		if ( 'radio-text' === $type && isset( $attr['size'] ) ) {
+			$radio_text_size = $attr['size'];
+
+			if ( $radio_text_size ) {
+				unset( $attr['size'] );
+			}
+		}
+
 		/* Placeholder */
 
 		$placeholder = $placeholder ? 'placeholder="' . esc_attr( $placeholder ) . '"' : '';
@@ -525,11 +537,13 @@ class Field {
 		/* Required */
 
 		$req_attr = '';
+		$req_icon = '';
 		$attr     = Utils::get_attr_as_str(
 			$attr,
-			function( $a, $v ) use ( &$req_attr ) {
+			function( $a, $v ) use ( &$req_attr, &$req_icon ) {
 				if ( 'aria-required' === $a && 'true' === $v ) {
 					$req_attr = ' data-required';
+					$req_icon = '<span data-required-icon aria-hidden="true"></span>';
 				}
 			}
 		);
@@ -559,14 +573,14 @@ class Field {
 			if ( $checkbox_radio ) {
 				$label = (
 					"<label id='$label_id' for='" . esc_attr( $id ) . "'$req_attr>" .
-						"<span data-label class='$label_class'><span>$label</span></span>" .
+						"<span data-label class='$label_class'><span>$label$req_icon</span></span>" .
 						'<span data-control data-type="' . $type . '"></span>' .
 					'</label>'
 				);
 			} else {
 				$label = (
 					"<label data-label id='$label_id' class='$label_class' for='" . esc_attr( $id ) . "'$req_attr>" .
-						"<span>$label</span>" .
+						"<span>$label$req_icon</span>" .
 					'</label>'
 				);
 			}
@@ -798,17 +812,23 @@ class Field {
 		}
 
 		if ( 'radio-text' === $type ) {
+			$radio_text_attr = [
+				'aria-label'  => $label_text,
+				'disabled'    => '',
+				'data-enable' => $id,
+			];
+
+			if ( $radio_text_size ) {
+				$radio_text_attr['size'] = $radio_text_size;
+			}
+
 			$output .= self::render_field(
 				[
 					'id'    => uniqid(),
 					'name'  => $name . '_text',
 					'type'  => 'text',
 					'class' => 'js-conditional',
-					'attr'  => [
-						'aria-label'  => $label_text,
-						'disabled'    => '',
-						'data-enable' => $id,
-					],
+					'attr'  => $radio_text_attr,
 				],
 				$output
 			);